added mst functionality for the mst in doStpPortTask() function

+++ cfgmgr/stpmgr.h
@@ -290,7 +290,7 @@ private:
     void processStpMstInstPortAttr(const std::string op, uint16_t mst_id, const std::string intfName,
                                        std::vector<FieldValueTuple>&tupEntry);
     std::vector<uint16_t> parseVlanList(const std::string &vlanStr);
-    void updateVlanInstanceMap(int instance, const std::vector<int>&newVlanList, bool operation);
+    void updateVlanInstanceMap(int instance, const std::vector<uint16_t>&newVlanList, bool operation);
     bool isInstanceMapped(uint16_t instance);
 
 };
